% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clsuter_result.R
\name{get_cluster_result}
\alias{get_cluster_result}
\title{Obtain the clustering result}
\usage{
get_cluster_result(test_exp, ensembleResultLst, resultLst, cls_num)
}
\arguments{
\item{test_exp}{The pre-processed input data}

\item{ensembleResultLst}{This can be obtained from the automatically saved .RData file in your working directory.}

\item{resultLst}{This can be obtained from the automatically saved .RData file in your working directory.}

\item{cls_num}{The desired number of clusters to view (value > 10)}
}
\description{
This function obtains the clustering result from reCAT, bayes and mean scores, and the time series ordering.
}
\examples{
test_exp <- get_test_exp(data)
data_ordIndex <- get_ordIndex(test_exp, 2)
# Load the following .RData file from your local working directory, it is automatically generated when you perform the function 'get_ordIndex'
load("../data/bestEnsembleComplexTSP 10 - N  .RData")
cls_result = get_cluster_result(test_exp = test_exp, ensembleResultLst = ensembleResultLst,
                               resultLst = resultLst, cls_num = 20)
}
