% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_score}
\alias{get_score}
\title{Bayes scores and mean scores}
\usage{
get_score(talentDE)
}
\arguments{
\item{data}{The data with cycling genes, m genes by n cells}
}
\description{
This function produces the bayes scores (G1, S, G2/M) and Mean scores (G1, S, G2, M, G1/S, G2/M) over the dataset. The output can then be used to classify different cycle stages using HMM.
}
\examples{
test_exp <- get_test_exp(data)
get_score(t(test_exp))
}
\keyword{get_test_exp}
