% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start.R
\name{get_start}
\alias{get_start}
\title{Determining the starting cell}
\usage{
get_start(
  bayes_score,
  mean_score,
  ordIndex,
  cls_num,
  rdata = NULL,
  nthread = 3
)
}
\arguments{
\item{bayes_score}{Obtained from get_score function}

\item{mean_score}{Obtained from get_score function}

\item{ordIndex}{Obtained from get_ordIndex function}

\item{cls_num}{The number of cell cycle stages that you want to segment, ie 3 - G1, S, G2/M.}

\item{rdata}{This sets the region of each cycle stage manually by observing the scores. Usage: rdata = t(data.frame(c(2,5), c(10,15), c(20,25))) corresponding to G1: 2-5, S: 10-15 and G2/M: 20-25. If NULL, it will prompt you to enter the region values such as '2 5', please enter twice to confirm each line of input.}

\item{nthreads}{The number of threads. Defaults to 3}
}
\description{
This function is used to find the starting cell of the cycle
}
\examples{
test_exp <- get_test_exp(data)
data_ordIndex <- get_ordIndex(test_exp, 2)
score_results <- get_score(t(test_exp))
rdata = t(data.frame(c(2,5), c(10,15), c(20,25)))
data_start <- get_start(score_results$bayes_score, score_results$mean_score, data_ordIndex, cls_num = 3, rdata,  nthread = 3)
}
