% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_exp.R
\name{get_test_exp}
\alias{get_test_exp}
\title{Pre-processing the gene expression}
\usage{
get_test_exp(data)
}
\arguments{
\item{data}{The gene expression data input, m genes by n cells}
}
\description{
This function is used to get the correct input to reCAT and automatically selects the cycling genes from the Cyclebase
}
\examples{
get_test_exp(data) 
}
