% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bayes}
\alias{plot_bayes}
\title{Visualizing Bayes scores}
\usage{
plot_bayes(
  bayes_score,
  ordIndex,
  cls_result = NULL,
  cls_ord = NULL,
  colorbar = 0
)
}
\arguments{
\item{bayes_score}{Obtained from get_score function}

\item{ordIndex}{Obtained from get_ordIndex function}

\item{cls_result}{The result obtained from HMM segmentation. If null, plot is produced without colorbar representing the segments. Defaults to NULL}

\item{colorbar}{Whether to plot colorbar (=1) or not (=0). Defaults to 0}

\item{myord}{= The determined start and direction of the time series. e.g. myord=c(4:1, 295:5) denotes that the starting cell is from 4, and orders backwards. Defaults to NULL}
}
\description{
This function is used to visualize the Bayes scores of the dataset along the pseudo time series
}
\examples{
test_exp <- get_test_exp(data)
data_ordIndex <- get_ordIndex(test_exp, 2)
score_results <- get_score(t(test_exp))
myord = c(4:1, 295:5)
hmm_result <- get_hmm_order(bayes_score = score_results$bayes_score,
                           mean_score = score_results$mean_score,
                           ordIndex = data_ordIndex, cls_num = 3, myord = myord)
plot_bayes(score_results$bayes_score, data_ordIndex, cls_result = hmm_result, myord = myord, colorbar = 0)
}
