% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reCAT.R
\docType{package}
\name{reCAT-package}
\alias{reCAT}
\alias{reCAT-package}
\title{reCAT: Reconstructing Cell Cycle Pseudo Time-Series via Single-Cell Transcriptome Data}
\description{
reCAT is a modelling framework for unsynchronized single-cell transcriptome data that can reconstruct a high-resolution cell cycle time-series. It is mainly based on traveling salesman problem (TSP), Gaussian mixture model (GMM) and hidden Markov model (HMM). We developed an R based software package which is easy to use. The performance is relatively accurate and stable. Thanks for using!
}
\references{
Liu, Z., Lou, H., Xie, K., Wang, H., Chen, N., Aparicio, O.M., Zhang, M.Q., Jiang, R. and Chen, T., 2017. Reconstructing cell cycle pseudo time-series via single-cell transcriptome data. Nature communications, 8(1), pp.1-9.
}
\author{
\strong{Maintainer}: Zehua Liu \email{zehualiuf@hotmail.com}

Authors:
\itemize{
  \item Huazhe Lou [contributor]
  \item Ken Xie [contributor]
  \item Hao Wang [contributor]
}

}
\keyword{internal}
